refactor: move `oui` under `discover` package + move scanners under dedicated package
- preparation for making a module out of the `discovery` package
