+++ internal/core/discovery/scanners/arp/darwin_cache_reader_stub.go
