+++ internal/core/discovery/scanners/arp/linux_cache_reader_stub.go
