+++ internal/core/discovery/scanners/arp/windows_cache_reader_stub.go
