+++ internal/core/discovery/scanners/ssdp/ssdp.go
