+++ internal/core/discovery/scanners/ssdp/ssdp_test.go
