+++ internal/core/engine_builder.go
@@ -3,10 +3,10 @@ package core
 import (
 	"github.com/ramonvermeulen/whosthere/internal/core/config"
 	"github.com/ramonvermeulen/whosthere/internal/core/discovery"
-	"github.com/ramonvermeulen/whosthere/internal/core/discovery/arp"
-	"github.com/ramonvermeulen/whosthere/internal/core/discovery/mdns"
-	"github.com/ramonvermeulen/whosthere/internal/core/discovery/ssdp"
-	"github.com/ramonvermeulen/whosthere/internal/core/oui"
+	"github.com/ramonvermeulen/whosthere/internal/core/discovery/oui"
+	"github.com/ramonvermeulen/whosthere/internal/core/discovery/scanners/arp"
+	"github.com/ramonvermeulen/whosthere/internal/core/discovery/scanners/mdns"
+	"github.com/ramonvermeulen/whosthere/internal/core/discovery/scanners/ssdp"
 )
 
 func BuildEngine(iface *discovery.InterfaceInfo, ouiDB *oui.Registry, cfg *config.Config) *discovery.Engine {
